move pkg/util/rand
