+++ pkg/api/v1/BUILD
@@ -37,7 +37,6 @@ go_library(
         "//pkg/util:go_default_library",
         "//pkg/util/intstr:go_default_library",
         "//pkg/util/parsers:go_default_library",
-        "//pkg/util/rand:go_default_library",
         "//vendor:github.com/gogo/protobuf/proto",
         "//vendor:github.com/gogo/protobuf/sortkeys",
         "//vendor:github.com/ugorji/go/codec",
@@ -49,6 +48,7 @@ go_library(
         "//vendor:k8s.io/apimachinery/pkg/runtime/schema",
         "//vendor:k8s.io/apimachinery/pkg/selection",
         "//vendor:k8s.io/apimachinery/pkg/types",
+        "//vendor:k8s.io/apimachinery/pkg/util/rand",
         "//vendor:k8s.io/apimachinery/pkg/util/sets",
         "//vendor:k8s.io/apimachinery/pkg/util/validation/field",
     ],
