+++ pkg/cloudprovider/providers/photon/BUILD
@@ -29,8 +29,8 @@ go_test(
     tags = ["automanaged"],
     deps = [
         "//pkg/cloudprovider:go_default_library",
-        "//pkg/util/rand:go_default_library",
         "//vendor:k8s.io/apimachinery/pkg/types",
+        "//vendor:k8s.io/apimachinery/pkg/util/rand",
     ],
 )
