+++ pkg/kubelet/BUILD
@@ -184,7 +184,6 @@ go_test(
         "//pkg/util/clock:go_default_library",
         "//pkg/util/flowcontrol:go_default_library",
         "//pkg/util/mount:go_default_library",
-        "//pkg/util/rand:go_default_library",
         "//pkg/util/strategicpatch:go_default_library",
         "//pkg/util/testing:go_default_library",
         "//pkg/util/uuid:go_default_library",
@@ -203,6 +202,7 @@ go_test(
         "//vendor:k8s.io/apimachinery/pkg/runtime",
         "//vendor:k8s.io/apimachinery/pkg/types",
         "//vendor:k8s.io/apimachinery/pkg/util/diff",
+        "//vendor:k8s.io/apimachinery/pkg/util/rand",
         "//vendor:k8s.io/apimachinery/pkg/util/runtime",
         "//vendor:k8s.io/apimachinery/pkg/util/sets",
         "//vendor:k8s.io/apimachinery/pkg/util/wait",
