+++ pkg/util/rand/BUILD
@@ -5,19 +5,11 @@ licenses(["notice"])
 load(
     "@io_bazel_rules_go//go:def.bzl",
     "go_library",
-    "go_test",
 )
 
 go_library(
     name = "go_default_library",
-    srcs = ["rand.go"],
-    tags = ["automanaged"],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["rand_test.go"],
-    library = ":go_default_library",
+    srcs = ["doc.go"],
     tags = ["automanaged"],
 )
