+++ test/e2e/framework/BUILD
@@ -75,7 +75,6 @@ go_library(
         "//pkg/util/exec:go_default_library",
         "//pkg/util/intstr:go_default_library",
         "//pkg/util/labels:go_default_library",
-        "//pkg/util/rand:go_default_library",
         "//pkg/util/system:go_default_library",
         "//pkg/util/uuid:go_default_library",
         "//pkg/util/version:go_default_library",
@@ -106,6 +105,7 @@ go_library(
         "//vendor:k8s.io/apimachinery/pkg/types",
         "//vendor:k8s.io/apimachinery/pkg/util/errors",
         "//vendor:k8s.io/apimachinery/pkg/util/net",
+        "//vendor:k8s.io/apimachinery/pkg/util/rand",
         "//vendor:k8s.io/apimachinery/pkg/util/runtime",
         "//vendor:k8s.io/apimachinery/pkg/util/sets",
         "//vendor:k8s.io/apimachinery/pkg/util/validation",
