+++ test/images/logs-generator/logs_generator.go
@@ -22,7 +22,7 @@ import (
 	"time"
 
 	"github.com/golang/glog"
-	"k8s.io/kubernetes/pkg/util/rand"
+	"k8s.io/apimachinery/pkg/util/rand"
 )
 
 var (
